+++ tests/conftest.py
@@ -1,11 +1,14 @@
 from __future__ import annotations
 
+import os
 import sys
 from contextlib import contextmanager
 from typing import Callable
 
 import pytest
 from click.testing import CliRunner
+from nbmake.pytest_items import NotebookItem
+from pytask import storage
 
 
 class SysPathsSnapshot:
@@ -62,6 +65,7 @@ def _restore_sys_path_and_module_after_test_execution():
 class CustomCliRunner(CliRunner):
     def invoke(self, *args, **kwargs):
         """Restore sys.path and sys.modules after an invocation."""
+        storage.create()
         with restore_sys_path_and_module_after_test_execution():
             return super().invoke(*args, **kwargs)
 
@@ -69,3 +73,11 @@ class CustomCliRunner(CliRunner):
 @pytest.fixture()
 def runner():
     return CustomCliRunner()
+
+
+def pytest_collection_modifyitems(session, config, items) -> None:  # noqa: ARG001
+    """Add markers to Jupyter notebook tests."""
+    if sys.platform == "darwin" and "CI" in os.environ:  # pragma: no cover
+        for item in items:
+            if isinstance(item, NotebookItem):
+                item.add_marker(pytest.mark.xfail(reason="Fails regularly on MacOS"))