+++ CHANGES.md
@@ -11,6 +11,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask-parallel) and
 - {pull}`75` updates the release strategy.
 - {pull}`79` add tests for Jupyter and fix parallelization with `PythonNode`s.
 - {pull}`80` adds support for partialed functions.
+- {pull}`82` fixes testing with pytask v0.4.5.
 
 ## 0.4.0 - 2023-10-07